+++ modules/Discipline/MakeReferral.php
@@ -154,13 +154,6 @@ if ( ! $_REQUEST['modfunc'] )
 
 		PopTable( 'header', ProgramTitle() );
 
-		$categories_RET = DBGet( "SELECT df.ID,df.DATA_TYPE,du.TITLE,du.SELECT_OPTIONS
-			FROM DISCIPLINE_FIELDS df,DISCIPLINE_FIELD_USAGE du
-			WHERE du.SYEAR='" . UserSyear() . "'
-			AND du.SCHOOL_ID='" . UserSchool() . "'
-			AND du.DISCIPLINE_FIELD_ID=df.ID
-			ORDER BY du.SORT_ORDER" );
-
 		echo '<table class="width-100p"><tr><td>';
 
 		$users_RET = DBGet( "SELECT STAFF_ID," . DisplayNameSQL() . " AS FULL_NAME,
@@ -255,9 +248,16 @@ if ( ! $_REQUEST['modfunc'] )
 
 		echo '</td></tr></table></td></tr>';
 
+		$categories_RET = DBGet( "SELECT df.ID,df.DATA_TYPE,du.TITLE,du.SELECT_OPTIONS
+			FROM DISCIPLINE_FIELDS df,DISCIPLINE_FIELD_USAGE du
+			WHERE du.SYEAR='" . UserSyear() . "'
+			AND du.SCHOOL_ID='" . UserSchool() . "'
+			AND du.DISCIPLINE_FIELD_ID=df.ID
+			ORDER BY du.SORT_ORDER" );
+
 		foreach ( (array) $categories_RET as $category )
 		{
-			echo '<tr><td>' . ReferralInput( $category ) . '</td></tr>';
+			echo '<tr><td><hr />' . ReferralInput( $category ) . '</td></tr>';
 		}
 
 		echo '</table>';