+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes in 9.0
 --------------
 - CSS add length to previous meals select in DailyMenus.php
 - CSS FlatSIS fix calendar menu text wrapping in stylesheet.css
+- Add Export list button in TransactionsReport.php
 
 Changes in 8.9
 --------------