+++ CHANGES.md
@@ -11,6 +11,7 @@ Changes in 9.0
 - HTML add horizontal ruler before each category in MakeReferral.php
 - Fix SQL error when generating Schedule table with PHP8.1 in GetMP.php
 - Reorder PDF list columns to match Schedule columns in PrintSchedules.php
+- SQL order Schedule list by Course Title & Course Period Short Name in Schedule.php & PrintSchedules.php
 
 Changes in 8.9
 --------------