+++ CHANGES.md
@@ -13,6 +13,7 @@ Changes in 9.0
 - Reorder PDF list columns to match Schedule columns in PrintSchedules.php
 - SQL order Schedule list by Course Title & Course Period Short Name in Schedule.php & PrintSchedules.php
 - Fix SQL error more than one row returned by a subquery in Rollover.php
+- Fix update Course Period title when Short Name contains single quote in Courses.php
 
 Changes in 8.9
 --------------